api: refactor idle killer
This fixes a basic mistake I made assuming that the hub's main code in
the websocket loops continues to exist after the browser exits. So now
the idle killer daemon runs in a separate asyncio task.

This also includes a much-needed refactor of the TTY websocket code to
separate all the pod management into its own file.