tests: Fixes for tests not running GoTTY
The way that run.sh was checking to see if it is running a
guest or an account session was by seeing if the ACCOUNT_PATH was set.
That was clearly a fragile method. Now we check for USER_ID instead.

Also the `docker run` to run the vdwww container in the tests had an
extra line which stoppped the logs being output on failure.