[airtable] Implement basic loader
Without the meta API, we don't know the column definitions of the
table, so they need to be populated from the data. Here the columns
are populated from the first returned record, but this is not
sufficient since an empty value for a column in a record will not
be present at all. Later records need to be able to add missing
columns.